pyproject.toml
[project]
name = "python-uv-google-cloud-bigquery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["google-cloud-bigquery>=3.34.0", "gpiozero>=2.0.1"]

[[tool.mypy.overrides]]
module = ["gpiozero", "gpiozero.pins.pigpio"]
ignore_missing_imports = true
